Move test to `complete-onboarding-wizard`
