+++ plugins/woocommerce/tests/e2e/specs/activate-and-setup/complete-onboarding-wizard.test.js
@@ -4,6 +4,7 @@ const {
 	testDifferentStoreCurrenciesWCPay,
 	testSubscriptionsInclusion,
 	testBusinessDetailsForm,
+	testAdminHomescreenWelcomeModal,
 } = require( '@woocommerce/admin-e2e-tests' );
 const { withRestApi, IS_RETEST_MODE } = require( '@woocommerce/e2e-utils' );
 
@@ -17,3 +18,4 @@ testSelectiveBundleWCPay();
 testDifferentStoreCurrenciesWCPay();
 testSubscriptionsInclusion();
 testBusinessDetailsForm();
+testAdminHomescreenWelcomeModal();
